Make WebGestureEvent coords and device private
This CL changes WebGestureEvent coordinates to private and float
to match other Web* class e.g. WebMouseEvent.

x, y -> position_in_widget_
global_x, global_y -> position_in_screen_

This also changes source_device to private source_device_.
Add set source_device_ in constructor, and add setter and getter.
SetSourceDevice will be eliminated in future CL (should be set in constructor)

Bug: 817104
Cq-Include-Trybots: luci.chromium.try:linux_optional_gpu_tests_rel;master.tryserver.blink:linux_trusty_blink_rel;master.tryserver.chromium.android:android_optional_gpu_tests_rel;master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2;master.tryserver.chromium.linux:linux_optional_gpu_tests_rel;master.tryserver.chromium.linux:linux_vr;master.tryserver.chromium.mac:mac_optional_gpu_tests_rel;master.tryserver.chromium.win:win_optional_gpu_tests_rel
Reviewed-on: https://chromium-review.googlesource.com/953655
Cr-Commit-Position: refs/heads/master@{#544619}

+++ content/browser/renderer_host/input/web_input_event_builders_android.h
@@ -50,8 +50,8 @@ class WebGestureEventBuilder {
  public:
   static blink::WebGestureEvent Build(blink::WebInputEvent::Type type,
                                       double time_sec,
-                                      int x,
-                                      int y);
+                                      float x,
+                                      float y);
 };
 
 }  // namespace content
